Separate testing helper functions into multiple files
